Update instantiate tests
