+++ src/tests.rs
@@ -1,4 +1,5 @@
 pub mod execute;
+pub mod instantiate;
 pub mod query;
 pub mod test_constants;
 pub mod test_setup_utils;
