+++ src/tests/instantiate.rs
@@ -1,0 +1,1 @@
+mod instantiate_tests;
